inject into controllet gateway, and emit dialogs event when create dialog

+++ back/src/dialog/dialog.controller.ts
@@ -1,5 +1,5 @@
-import { Body, Controller, Delete, Get, NotFoundException, Param, Post, Query, UseGuards, Request } from "@nestjs/common";
-import { ApiBadRequestResponse, ApiCreatedResponse, ApiNotFoundResponse, ApiOkResponse, ApiResponse, ApiTags, ApiUnauthorizedResponse } from "@nestjs/swagger";
+import { Body, Controller, Delete, Get, NotFoundException, Param, Post, Query, UseGuards, Request, BadRequestException, ForbiddenException } from "@nestjs/common";
+import { ApiBadRequestResponse, ApiCreatedResponse, ApiForbiddenResponse, ApiNotFoundResponse, ApiOkResponse, ApiResponse, ApiTags, ApiUnauthorizedResponse } from "@nestjs/swagger";
 import { CreateMessageDto } from "./../message/dto/create-message.dto";
 import { MessageService } from "./../message/message.service";
 import UserService from "./../user/user.service";
@@ -7,13 +7,14 @@ import { IdValidationPipe } from "./../pipe/id-validation.pipe";
 import { DialogService } from "./dialog.sevice";
 import { CreateDialogDto } from "./dto/create-dialog.dto";
 import { Dialog } from "./entities/dialog.entity";
-import { DIALOG_NOT_FOUND } from "./const";
+import { DIALOG_CANT_DELETE, DIALOG_CREATE_ERROR, DIALOG_NOT_FOUND } from "./const";
 import { ID_VALIDATION_ERROR } from "~/pipe/id-validation.contstants";
 import { Message } from "~/message/entities/message.entity";
 import { JwtAuthGuard } from "~/auth/guard/jwt-auth.guard";
 import {  IReqAuth } from "~/auth/interface/jwt.interface";
 import { USER_NOT_FOUND } from "~/user/const";
 import { FailedRequestResponse, TypesFailedResponse } from "~/types";
+import { DialogGateway } from "./dialog.gateway";
 
 @UseGuards(JwtAuthGuard)
 @ApiUnauthorizedResponse({
@@ -28,7 +29,8 @@ import { FailedRequestResponse, TypesFailedResponse } from "~/types";
 @Controller('dialog')
 export class DialogController {
     constructor(
-        private readonly dialogService: DialogService
+        private readonly dialogService: DialogService,
+        private readonly dialogGateway: DialogGateway
     ) {
 
     }
@@ -53,7 +55,13 @@ export class DialogController {
 
         await this.dialogService.addMessage(dialog._id, messageDto);
 
-        return dialog
+        const dialogActual = await this.dialogService.find(dialog._id);
+        
+        if(!dialogActual) throw new BadRequestException(DIALOG_CREATE_ERROR);
+
+        this.dialogGateway.dialogsEmit(dialogActual);
+
+        return dialog;
     }
 
 
@@ -97,8 +105,18 @@ export class DialogController {
     @ApiBadRequestResponse({
         description: ID_VALIDATION_ERROR
     })
+    @ApiForbiddenResponse({
+        description: DIALOG_CANT_DELETE
+    })
     @Delete(':id')
-    async deleteDialog(@Param('id', IdValidationPipe) dialogId: string) {
+    async deleteDialog(
+        @Request() req: IReqAuth,
+        @Param('id', IdValidationPipe) dialogId: string
+    ) {
+        const isUserMembers = Boolean(await this.dialogService.dialogWhereUserMember(dialogId, req.user._id));
+
+        if(!isUserMembers) throw new ForbiddenException(DIALOG_CANT_DELETE)
+
         await this.dialogService.deleteDialog(dialogId);
         return;
     }
